Mise en place d'un routing côté client
