+++ index.html
@@ -10,13 +10,18 @@
     <meta name="description" content=" ">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
+    <link id="especes-data" rel="preload" as="fetch" href="./data/liste_especes.csv" crossorigin="anonymous">
+    <link id="activites-data" rel="preload" as="fetch" href="./data/activités.csv" crossorigin="anonymous">
+    <link id="methodes-data" rel="preload" as="fetch" href="./data/méthodes.csv" crossorigin="anonymous">
+    <link id="transports-data" rel="preload" as="fetch" href="./data/transports.csv" crossorigin="anonymous">
+
     <link crossorigin="anonymous" rel="stylesheet" href="./style/bootstrap-reset-4.6.2.css">
     <link crossorigin="anonymous" rel="stylesheet" href="./style/better-css-defaults.css">
     <link crossorigin="anonymous" rel="stylesheet" href="./style/layout.css">
 
-    <link crossorigin="anonymous" rel="stylesheet" href="./build/bundle-suivi.css">
+    <link crossorigin="anonymous" rel="stylesheet" href="./build/rollup-bundle-pitchou.css">
     
-    <script src="./build/bundle-suivi.js" type="module" crossorigin="anonymous"></script>
+    <script src="./build/rollup-bundle-pitchou.js" type="module" crossorigin="anonymous"></script>
 </head>
 <body>
     <main>
