--- saisie-especes.html
@@ -1,31 +1,0 @@
-<!doctype html>
-<html lang="fr">
-<head>
-    <meta charset="utf-8">
-    <meta name="referrer" content="no-referrer">
-    <link rel="icon" href="data:,">
-
-    <title>Saisie espèces protégées</title>
-    
-    <meta name="description" content=" ">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link id="especes-data" rel="preload" as="fetch" href="./data/liste_especes.csv" crossorigin="anonymous">
-    <link id="activites-data" rel="preload" as="fetch" href="./data/activités.csv" crossorigin="anonymous">
-    <link id="methodes-data" rel="preload" as="fetch" href="./data/méthodes.csv" crossorigin="anonymous">
-    <link id="transports-data" rel="preload" as="fetch" href="./data/transports.csv" crossorigin="anonymous">
-
-    <link crossorigin="anonymous" rel="stylesheet" href="./style/bootstrap-reset-4.6.2.css">
-    <link crossorigin="anonymous" rel="stylesheet" href="./style/better-css-defaults.css">
-    <link crossorigin="anonymous" rel="stylesheet" href="./style/layout.css">
-
-    <link crossorigin="anonymous" rel="stylesheet" href="./build/bundle-saisie-especes.css">
-    
-    <script src="./build/bundle-saisie-especes.js" type="module" crossorigin="anonymous"></script>
-</head>
-<body>
-    <main>
-        <section class="svelte-main"></section>
-    </main>
-</body>
-</html>
